+++ src/extension_commands.ml
@@ -489,7 +489,7 @@ end
 module Copy_type_under_cursor = struct
   let extension_name = "Copy Type Under Cursor"
 
-  let ocaml_lsp_doesnt_support_type_enclosing instance ocaml_lsp =
+  let ocaml_lsp_doesnt_support_type_selection instance ocaml_lsp =
     match
       Ocaml_lsp.is_version_up_to_date
         ocaml_lsp
@@ -510,8 +510,8 @@ module Copy_type_under_cursor = struct
     Custom_requests.(
       send_request
         client
-        Type_enclosing.request
-        (Type_enclosing.make
+        Type_selection.request
+        (Type_selection.make
            ~uri
            ~at:(`Range (Selection.to_range selection))
            ~index:0
@@ -531,13 +531,13 @@ module Copy_type_under_cursor = struct
         | Some text_editor ->
           (match Extension_instance.lsp_client instance with
            | None -> show_message `Warn "ocamllsp is not running" |> Promise.return
-           | Some (_, ocaml_lsp) when not (Ocaml_lsp.can_handle_type_enclosing ocaml_lsp)
+           | Some (_, ocaml_lsp) when not (Ocaml_lsp.can_handle_type_selection ocaml_lsp)
              ->
-             ocaml_lsp_doesnt_support_type_enclosing instance ocaml_lsp |> Promise.return
+             ocaml_lsp_doesnt_support_type_selection instance ocaml_lsp |> Promise.return
            | Some (client, _) ->
              let clipboard = Env.clipboard () in
              let open Promise.Syntax in
-             let* Custom_requests.Type_enclosing.{ type_; _ } =
+             let* Custom_requests.Type_selection.{ type_; _ } =
                get_enclosings text_editor client
              in
              if String.equal type_ "<no information>"
@@ -1132,6 +1132,13 @@ module Navigate_holes = struct
     command Extension_consts.Commands.navigate_typed_holes handler
   ;;
 end
+
+let _type_selection =
+  let open Type_selection in
+  command Extension_consts.Commands.type_selection handler |> ignore;
+  command Extension_consts.Commands.type_previous_selection previous_handler |> ignore;
+  command Extension_consts.Commands.augment_selection_type_verbosity verbosity_handler
+;;
 
 let register extension instance = function
   | Command { id; handler } ->